Update greeting.py
Trying to fully keep track of joined users for deciding when to (not) greet depending on their presence in the channel.
This is meant as an improvement over https://github.com/apertium/phenny/pull/428 to address https://github.com/apertium/phenny/issues/427.
I use input.group() since it was mentioned that input.args may not be passed correctly, but no idea if THAT will work instead.
It's untested as my Phenny bot lacks the database stuff that begiak uses in this module.
I don't know if this is the proper way to make a PR, as I'm still unfamiliar with Git and GitHub. Sorry in advance.